feat: wip on refactoring
