+++ nvim/.config/nvim/init.lua
@@ -1,21 +1,12 @@
---[[ init.lua ]] for _, source in ipairs {"utils"} do
-    local status_ok, fault = pcall(require, source)
-    if not status_ok then
-        vim.api.nvim_err_writeln("Failed to load " .. source .. "\n\n" .. fault)
-    end
+--[[ init.lua ]]
+for _, source in ipairs { "utils" } do
+  local status_ok, fault = pcall(require, source)
+  if not status_ok then
+    vim.api.nvim_err_writeln("Failed to load " .. source .. "\n\n" .. fault)
+  end
 end
 
-local lazypath = vim.fn.stdpath("data") .. "/lazy/lazy.nvim"
-if not vim.loop.fs_stat(lazypath) then
-    vim.fn.system({
-        "git", "clone", "--filter=blob:none",
-        "https://github.com/folke/lazy.nvim.git", "--branch=stable", -- latest stable release
-        lazypath
-    })
-end
-vim.opt.rtp:prepend(lazypath)
-
-require('vars') -- Variables
-require('opts') -- Options
-require('plugins') -- Pluginm
-require('mappings') -- Keymaps
+require('config.vars')
+require('config.options')
+require('config.plugins')
+require('config.mappings')
