+++ nvim/.config/nvim/lua/config/plugins.lua
@@ -1,19 +1,19 @@
 -- Packer file
 -- require("plugins.packer")
-require("plugins.lazy")
-require("plugins.theme")
+require("config.lazy")
+require("config.theme")
 
 
 -- If Nvim is not running under VSCode enable those extensions
 if not neovim.is_vscode() then
   if vim.g.lsp_setup_ready == nil then
     vim.g.lsp_setup_ready = true
-    require('plugins.cmp')
+    require('config.cmp')
 
     -- Plugin custom configs
-    neovim.load_folder("plugins.configs")
+    neovim.load_folder("configs")
 
     -- Lsp configs only pure nvim
-    neovim.load_folder("plugins.lsp")
+    neovim.load_folder("lsp")
   end
 end
