+++ nvim/.config/nvim/lua/config/vars.lua
