+++ nvim/.config/nvim/lua/configs/colorizer.lua
