+++ nvim/.config/nvim/lua/configs/devicons.lua
