+++ nvim/.config/nvim/lua/configs/mason.lua
