+++ nvim/.config/nvim/lua/lsp/bash.lua
